% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_history.R
\name{animate_population}
\alias{animate_population}
\title{Animation of particles motion for 1D and 2D function minimization}
\usage{
animate_population(minimizer_result, n_points = 1000, ...)
}
\arguments{
\item{minimizer_result}{an object of class \code{MinimizationResult}.}

\item{n_points}{number of points used to draw the objective function.}

\item{...}{additional arguments, such as graphical parameters (see \link[graphics]{contour}).}
}
\value{
???
}
\description{
XXXX
}
\examples{
library(EmiR)

x1 <- parameter("x1", -32.768, 32.768)
x2 <- parameter("x2", -32.768, 32.768)
l <- list(x1, x2)

config <- config_ps(iterations = 250, n_particles = 150)
ps <- minimize_ps(obj_func = ackley_func,
                  parameters = l,
                  config = config)
plot_history(ps)
}
