% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__ABC.R
\name{config_abc}
\alias{config_abc}
\title{Create a configuration object for the ABC minimizer}
\usage{
config_abc(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  employed_frac = 0.5,
  n_scout = 1
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of bats.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{initial_loudness}{???}

\item{alpha}{???}

\item{initial_pulse_rate}{???}

\item{gamma}{???}

\item{freq_min}{???}

\item{freq_max}{???}
}
\value{
\code{config_abc} returns an object of class \code{ABCConfig}.
}
\description{
Create a configuration object for the BAT minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the number of particles (parameter \code{n_particles}) have
to be provided.
}
