% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__BAT.R
\name{config_bat}
\alias{config_bat}
\title{Create a configuration object for the BAT minimizer}
\usage{
config_bat(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  initial_loudness = 1.5,
  alpha = 0.9,
  initial_pulse_rate = 0.5,
  gamma = 0.9,
  freq_min = 0,
  freq_max = 2
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of bats.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{initial_loudness}{???}

\item{alpha}{???}

\item{initial_pulse_rate}{???}

\item{gamma}{???}

\item{freq_min}{???}

\item{freq_max}{???}
}
\value{
\code{config_bat} returns an object of class \code{BATConfig}.
}
\description{
Create a configuration object for the BAT minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the number of particles (parameter \code{n_particles}) have
to be provided.
}
