% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RInterface_CS.R
\name{config_CS}
\alias{config_CS}
\title{Create a configuration object for the CS minimizer}
\usage{
config_CS(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  mutation_rate = 0.25
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of ???}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{mutation_rate}{???.}
}
\value{
\code{config_CS} returns an object of class \code{CSConfig}.
}
\description{
Create a configuration object for the CS minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the population size (parameter \code{population_size}) have
to be provided.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_CS(iterations = 250, n_particles = 100)
ps <- minimize_PS(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(ps)
}
