% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_Interface.R
\name{config_ga}
\alias{config_ga}
\title{Create a configuration object for the GA minimizer}
\usage{
config_ga(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  keep_fraction = 0.4,
  m_mutation_rate = 0.1,
  penalty_parameter = 10
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of chromosomes.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{keep_fraction}{???.}

\item{m_mutation_rate}{???.}

\item{penalty_parameter}{penalty parameter in constrained optimization.}
}
\value{
\code{config_ga} returns an object of class \code{GAConfig}.
}
\description{
Create a configuration object for the GA minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the number of chromosomes (parameter \code{population_size}) have
to be provided.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_ga(iterations = 250, n_particles = 100)
GA <- minimize_GA(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(GA)
}
