% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__GSA.R
\name{config_gsa}
\alias{config_gsa}
\title{Create a configuration object for the GSA minimizer}
\usage{
config_gsa(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  gravitation = 100,
  velocity_max = 0.01
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of planets.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{gravitation}{Gravitational constant}

\item{velocity_max}{Maximum velocity in \% of the range of parameters}
}
\value{
\code{config_gsa} returns an object of class \code{GSAConfig}.
}
\description{
Create a configuration object for the GSA minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the number of planets (parameter \code{population_size}) have
to be provided.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_gsa(iterations = 250, population_size = 100)
GSA <- minimize_gsa(obj_func = eggholder,
                  parameters = l,
                  config = config)
print(GSA)
}
