% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__HS.R
\name{config_hs}
\alias{config_hs}
\title{Create a configuration object for the Harmony Search algorithm}
\usage{
config_hs(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  considering_rate = 0.5,
  adjusting_rate = 0.5,
  distance_bandwidth = 0.1
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of solutions in the harmony memory.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{considering_rate}{probability for each component of a newly generated solution
to be recalled from the harmony memory.}

\item{adjusting_rate}{probability of the pitch adjustment in case of a component
recalled from the harmony memory.}

\item{distance_bandwidth}{amplitude of the random pitch adjustment.}
}
\value{
\code{config_hs} returns an object of class \code{HSConfig}.
}
\description{
Create a configuration object for the Harmony Search algorithm (see
\link[EmiR]{minimize_hs} for more details on the algorithm). At minimum the number of iterations
(parameter \code{iterations}) and the number of particles (parameter \code{n_particles})
have to be provided.
}
