% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_Interface.R
\name{config_sa}
\alias{config_sa}
\title{Create a configuration object for the SA minimizer}
\usage{
config_sa(
  iterations,
  n_particles,
  iterations_same_cost = NULL,
  cognitive = 2,
  social = 2,
  dumping = 0.1,
  temperature = 50,
  penalty_parameter = 10
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{n_particles}{number of particles.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{cognitive}{???.}

\item{social}{???.}

\item{dumping}{???.}

\item{temperature}{temperature.}

\item{penalty_parameter}{penalty parameter in constrained optimization.}
}
\value{
\code{config_sa} returns an object of class \code{SAConfig}.
}
\description{
Create a configuration object for the SA minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the number of particles (parameter \code{n_particles}) have
to be provided.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_sa(iterations = 250, n_particles = 100)
SA <- minimize_SA(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(SA)
}
