% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERFACE__ConstrainedFunction.R
\name{constrained_function}
\alias{constrained_function}
\title{Constrained function for minimization}
\usage{
constrained_function(func, ...)
}
\arguments{
\item{func}{original objective function.}

\item{...}{one or more constraints of class \code{Constraint}. See \link[EmiR]{constraint}.}
}
\value{
\code{constrained_function} returns an object of class \code{ConstrainedFunction}.
}
\description{
Create a constrained function for minimization.
}
