% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERFACE__Constraint.R
\name{constraint}
\alias{constraint}
\title{Constraint for minimization}
\usage{
constraint(func, inequality)
}
\arguments{
\item{func}{function describing the constraint.}

\item{inequality}{inequality type. Possible values: \code{>}, \code{>=}, \code{<=}, \code{<}.}
}
\value{
\code{constraint} returns an object of class \code{Constraint}.
}
\description{
Create a constraint function for constrained optimization. Only inequality constraints are
supported.
}
\examples{
g1 <- function(x) 0.0193*x[3] - (x[1]*0.0625)
c1 <- constraint(g1, "<=")

}
