% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC__Freudenstein.R
\name{freudenstein_roth_func}
\alias{freudenstein_roth_func}
\title{Freudenstein Roth Function}
\usage{
freudenstein_roth_func(x)
}
\arguments{
\item{x}{numeric or complex vector.}
}
\value{
The value of the function.
}
\description{
\loadmathjax
Implementation of 2-dimensional Freudenstein Roth function.
}
\details{
On an 2-dimensional domain it is defined by

\mjdeqn{f(\vec{x}) = (x_{1} - 13 + ((5 - x_{2})x_{2} - 2)x_{2})^2 + (x_{1} - 29 + ((x_{2} + 1)x_{2} - 14)x_{2})^2}{f(x) = (x_{1} - 13 + ((5 - x_{2})x_{2} - 2)x_{2})^2 + (x_{1} - 29 + ((x_{2} + 1)x_{2} - 14)x_{2})^2}
and is usually evaluated on
\mjeqn{x_{i} \in \[ -10, 10 \]}{x_{i} in \[ -10, 10 \]}, for all
\mjeqn{i=1,2}{i=1,2}. The function has one global minimum at
\mjeqn{f(\vec{x}) = 0}{f(x) = 0} for \mjeqn{\vec{x} = \[ 5, 4 \]}{x = \[ 5, 4 \]}.
}
\references{
\insertRef{rao2019engineering}{EmiR}
}
