% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{data}
\name{G01InitPop}
\alias{G01InitPop}
\title{Data set for example G01}
\description{
This data set contains the initial positions for a population of size 20 to be used with
the example G01.
}
\keyword{data}
