% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Population.R
\name{get_population}
\alias{get_population}
\title{Get population positions}
\usage{
get_population(minimizer_result, iteration)
}
\arguments{
\item{minimizer_result}{an object of class \code{OptimizationResults} (see \link[EmiR]{OptimizationResults}).}

\item{iteration}{iteration number.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Return a \code{data.frame} with the position of all individuals in the population at
the specified iteration, from an object of class \code{OptimizationResults} produced with
the option \code{save_pop_history} set to \code{TRUE} (see \link[EmiR]{MinimizerOpts}).
}
