% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Minimize.R
\name{list_of_algorithms}
\alias{list_of_algorithms}
\title{Return the list of algorithms in EmiR}
\usage{
list_of_algorithms()
}
\value{
An object of class \code{data.frame}.
}
\description{
Return a \code{data.frame} with the ID, description and configuration function name of all the
algorithms implemented in EmiR.
}
