% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Minimize.R
\name{list_of_functions}
\alias{list_of_functions}
\title{Return the list of pre-defined functions in EmiR}
\usage{
list_of_functions()
}
\value{
An object of class \code{data.frame}.
}
\description{
Return a \code{data.frame} with function name, full name, and minimum and maximum number of parameters
accepted for all the pre-defined functions in EmiR.
}
