% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC__Miele.R
\name{miele_cantrell_func}
\alias{miele_cantrell_func}
\title{Miele Cantrell Function}
\usage{
miele_cantrell_func(x)
}
\arguments{
\item{x}{numeric or complex vector.}
}
\value{
The value of the function.
}
\description{
\loadmathjax
Implementation of 4-dimensional Miele Cantrell Function.
}
\details{
On an 4-dimensional domain it is defined by

\mjdeqn{f(\vec{x}) = \left(e^{-x_{1}} - x_{2} \right)^4 + 100(x_{2} - x_{3})^6 + \left(\tan(x_{3} - x_{4})\right)^4 + x_{1}^8}{f(x) = (e^{-x_{1}} - x_{2} )^4 + 100(x_{2} - x_{3})^6 + (tan(x_{3} - x_{4}))^4 + x_{1}^8}
and is usually evaluated on
\mjeqn{x_{i} \in \[ -2, 2 \]}{x_{i} in \[ -2, 2 \]}, for all
\mjeqn{i=1,...,4}{i=1,...,4}. The function has one global minimum at
\mjeqn{f(\vec{x}) = 0}{f(x) = 0} for \mjeqn{\vec{x} = \[ 0, 1, 1, 1 \]}{x = \[ 0, 1, 1, 1 \]}.
}
\references{
\insertRef{cragg1969study}{EmiR}
}
