% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__BAT.R
\name{minimize_bat}
\alias{minimize_bat}
\title{Bat algorithm for function minimization}
\usage{
minimize_bat(
  obj_func,
  parameters,
  config,
  constraints = NULL,
  silent_mode = FALSE
)
}
\arguments{
\item{obj_func}{objective function be minimized.}

\item{parameters}{list of parameters the function is minimized with respect to.
Parmeters are requested to be objects of class \code{Parameter}(see \link[EmiR]{parameter}).}

\item{config}{an object of class \code{BATConfig} with the configuration parameters
used by algorithm (see \link[EmiR]{config_bat}).}

\item{constraints}{list of constraints. Constraints are requested to be objects of
class \code{Constraint} (see \link[EmiR]{constraint}).}

\item{silent_mode}{if \code{TRUE} no messages are shown.}
}
\value{
\code{minimize_sa} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize an objective function, possibly subjected to inequality constraints, using
the Bat (BAT) algorithm.
}
\details{
\loadmathjax
Proposed in 2010 by Xin-She Yang, BAT is a heuristic algorithm that operates by
mimicking the echolocation behavior of bats to perform global optimization. It is
based on three rules, which mimic the natural behaviour of bats:
\enumerate{
\item All bats use echolocation to sense the distance.
\item Bats fly with the velocity \mjeqn{v_{i}}{v_{i}} at position \mjeqn{x_{i}}{x_{i}} and
and emit pulses of frequency \mjeqn{f_{i} \in [f_{min}, f_{max}]}{f_{i} \in \link{f_{min}, f_{max}}},
with rate \mjeqn{r_{i} \in [0, 1]}{r_{i} \in \link{0, 1}} and loudness
\mjeqn{A_{i} \in [A_{min}, A_{0}]}{A_{i} \in \link{A_{min}, A_{0}}}.
\item The loudness varies from a large (positive) \mjeqn{A_{0}}{A_{0}} to a minimum constant value \mjeqn{A_{min}}{A_{min}}.
}

BAT is an iterative algorithm which, after an initial random generation of bats (solutions)
in the search space, performs the following steps at each iteration:
\itemize{
\item \emph{Generation of a new solution}: XXX
}
}
\references{
\insertRef{yang2010new}{EmiR}
}
