% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__HS.R
\name{minimize_hs}
\alias{minimize_hs}
\title{Harmony Search algorithm for function minimization}
\usage{
minimize_hs(obj_func, parameters, config, constraints = NULL, ...)
}
\arguments{
\item{obj_func}{objective function be minimized.}

\item{parameters}{list of parameters the function is minimized with respect to.
Parmeters are requested to be objects of class \code{Parameter} (see \link[EmiR]{parameter}).}

\item{config}{an object of class \code{HSConfig} with the configuration parameters
used by algorithm (see \link[EmiR]{config_hs}).}

\item{constraints}{list of constraints. Constraints are requested to be objects of
class \code{Constraint} (see \link[EmiR]{constraint}).}

\item{silent_mode}{if \code{TRUE} no messages are shown.}
}
\value{
\code{minimize_hs} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize an objective function, possibly subjected to inequality constraints, using
the Harmony Search (HS) algorithm.
}
\details{
\loadmathjax
HS is a metaheuristic optimization algorithm, inspired by the process of improvisation
in music. When improvising, a musician tend to mix the following ideas, in the search
for a perfect harmony:
\enumerate{
\item playing a precise series of notes recalled from his or her memory;
\item playing some random notes;
\item playing something similar to what he or she already knows, by only changing
the pitch of some notes.
}

This approach can be also applied in the context of optimization problems. The starting
point of the HS algorithm is the random generation of possible solutions vector to the problem,
creating the so-called \emph{harmony memory}. At each iteration the following steps
are performed:
\enumerate{
\item \emph{Generation of a new solution}: values of the decision variables,
forming the new solution, are either chosen from the harmony memory, with a
probability \mjeqn{p_{cr}}{p_{cr}}, or randomly generated.
\mjeqn{p_{cr}}{p_{cr}} is called \emph{considering rate} and corresponds to the
parameter \code{considering_rate} in class \code{HSConfig} (see \link[EmiR]{config_hs}).
Each component obtained from the harmony memory is also pitch-adjusted with
a probability \mjeqn{p_{ar}}{p_{ar}}, called \emph{pitch adjusting rate}.
\mjeqn{p_{ar}}{p_{ar}} corresponds to the parameter \code{adjusting_rate}
in class \code{HSConfig}. The amount of the random pitch adjustment is
tuned by the parameter \code{distance_bandwidth} in class \code{HSConfig}.
\item \emph{Update of the harmony memory}: if the new solution improves the worst
one in the harmony memory (in terms of the cost of the objective function),
the latter is replaced by the new solution.
}

These steps are repeated until a stopping criterion is met (maximum number of
iterations, maximum number of iterations without improvements to the best solution, ...).
}
\references{
\insertRef{Lee2004}{EmiR}
}
