% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__IHS.R
\name{minimize_ihs}
\alias{minimize_ihs}
\title{Improved Harmony Search algorithm for function minimization}
\usage{
minimize_ihs(
  obj_func,
  parameters,
  config,
  constraints = NULL,
  silent_mode = FALSE
)
}
\arguments{
\item{obj_func}{objective function be minimized.}

\item{parameters}{list of parameters the function is minimized with respect to.
Parmeters are requested to be objects of class \code{Parameter}(see \link[EmiR]{parameter}).}

\item{config}{an object of class \code{IHSConfig} with the configuration parameters
used by algorithm (see \link[EmiR]{config_ihs}).}

\item{constraints}{list of constraints. Constraints are requested to be objects of
class \code{Constraint} (see \link[EmiR]{constraint}).}

\item{silent_mode}{if \code{TRUE} no messages are shown.}
}
\value{
\code{minimize_ihs} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize an objective function, possibly subjected to inequality constraints, using
the Improved Harmony Search (IHS) algorithm, as as proposed by Mahdavi \emph{et al.} in 2007.
}
\details{
\loadmathjax
IHS is a more general version of the Harmony Search algorithm, where the \emph{pitch adjusting rate}
(\mjeqn{p_{ar}}{p_{ar}}) and the amount of the random pitch adjustment (\mjeqn{A_{ar}}{A_{ar}}) evolve with the iterations, instead
of being fixed, as in the original algorithm (see \link[EmiR]{minimize_hs}).

The evolution of \mjeqn{p_{ar}}{p_{ar}} with the iteration \mjeqn{n}{n} is
\mjdeqn{p_{ar}(n) = p_{ar}^{min} + n \cdot \frac{p_{ar}^{max} - p_{ar}^{min}}{n_{T}},}{p_{ar}(n) = p_{ar,min} + n x (p_{ar,max} - p_{ar,min})/n_{T}}
where \mjeqn{p_{ar}^{min}}{p_{ar}^{min}} and \mjeqn{p_{ar}^{max}}{p_{ar}^{max}} are minimum and maximum
values for \mjeqn{p_{ar}}{p_{ar}}, set by the parameters \code{min_adjusting_rate} and
\code{max_adjusting_rate}, and \mjeqn{n_{T}}{n_{T}} is the total number of iterations (see \link[EmiR]{config_ihs}).

The evolution of \mjeqn{A_{ar}}{A_{ar}} with \mjeqn{n}{n} is
\mjdeqn{A_{ar}(n) = A_{ar}^{max} \exp\left[ \ln\left( \frac{A_{ar}^{min}}{A_{ar}^{max}} \right) \cdot \frac{n}{n_{T}} \right],}{A_{ar}(n) = A_{ar}^{max} exp\link{ ln( A_{ar}^{min}/A_{ar}^{max} ) x n/n_{T} }}
where \mjeqn{A_{ar}^{min}}{A_{ar}^{min}} and \mjeqn{A_{ar}^{max}}{A_{ar}^{max}} are minimum and maximum
values for \mjeqn{A_{ar}}{A_{ar}}, set by the parameters \code{min_distance_bandwidth} and \code{max_distance_bandwidth} (see \link[EmiR]{config_ihs}).
}
\references{
\insertRef{Mahdavi2007}{EmiR}
}
