% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__NEW.R
\name{minimize_new}
\alias{minimize_new}
\title{Artificial Bee Colony algorithm for function minimization}
\usage{
minimize_new(obj_func, parameters, config, constraints = NULL, ...)
}
\arguments{
\item{obj_func}{objective function be minimized.}

\item{parameters}{list of parameters composing the search space for the objective function.
Parmeters are requested to be objects of class \code{Parameter}(see \link[EmiR]{parameter}).}

\item{config}{an object of class \code{ABCConfig} with the configuration parameters
used by algorithm (see \link[EmiR]{config_new}).}

\item{constraints}{list of constraints. Constraints are requested to be objects of
class \code{Constraint} (see \link[EmiR]{constraint}).}

\item{...}{additional options (see \link[EmiR]{MinimizerOpts}).}
}
\value{
\code{minimize_new} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize an objective function, possibly subjected to inequality constraints, using
the Artificial Bee Colony (ABC) algorithm.
}
\references{
\insertRef{yang2010new}{EmiR}
}
