% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minimize_PS}
\alias{minimize_PS}
\title{Minimize a cost function using the Particle Swarm (PS) algorithm}
\usage{
minimize_PS(cost_function, parameters, config)
}
\arguments{
\item{cost_function}{cost function to be minimized.}

\item{parameters}{a list of objects of class "\code{Parameter}" the cost function is minimized with respect to.
See \link[EmiR]{parameter}.}

\item{config}{an object of class "\code{PSConfig}" with the configuration parameters
for the PS algorithm. See \link[EmiR]{config_PS}.}
}
\value{
\code{minimize_PS} returns an object of class "\code{MinimizationResult}".
}
\description{
Description of your package
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_PS(iterations = 250, n_particles = 100)
ps <- minimize_PS(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(ps)
}
