% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS__MinimizerOpts.R
\docType{class}
\name{MinimizerOpts-class}
\alias{MinimizerOpts-class}
\title{Options for EmiR optimization algorithms}
\arguments{
\item{silent_mode}{if \code{TRUE} no output to console is generated.}
}
\description{
Options for EmiR optimization algorithms
}
