% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RInterface_Parameter.R
\name{parameter}
\alias{parameter}
\title{Create XXX}
\usage{
parameter(name, min_val, max_val)
}
\arguments{
\item{name}{name.}

\item{min_val}{min_val.}

\item{max_val}{max_val.}
}
\value{
\code{parameter} returns an object of class "\code{Parameter}".
}
\description{
Create XXX
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_PS(iterations = 250, n_particles = 100)
ps <- minimize_PS(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(ps)
}
