% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTERFACE__Parameter.R
\name{parameters}
\alias{parameters}
\title{Set of parameters for minimization}
\usage{
parameters(values)
}
\arguments{
\item{values}{a \mjeqn{2 \times n}{2 x n} matrix or a \mjeqn{3 \times n}{3 x n} matrix.}
}
\value{
\code{parameters} returns a list of objects of class \code{Parameter}.
}
\description{
\loadmathjax
Create the set of parameters the objective function is minimized with respect to.
A \mjeqn{2 \times n}{2 x n} matrix or a \mjeqn{3 \times n}{3 x n} matrix, where the
first row is for the lower limits, the second one is for the upper limits, and the
(optional) third one is to specify if a parameter is constrained to be integer.
In case the third row is not provided, all the parameters are treated as continuous.
The name of each of the \mjeqn{n}{n} parameters is automatically generated and
it is of the form \mjeqn{xi}{xi}, where \mjeqn{i=1,...,n}{i=1,...,n}.
}
