% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Plot_history.R
\name{plot_history}
\alias{plot_history}
\title{Plot minimization history}
\usage{
plot_history(minimizer_result, ...)
}
\arguments{
\item{minimizer_result}{an object of class \code{MinimizationResult}.}

\item{...}{additional arguments, such as graphical parameters (see \link[base]{plot}).}
}
\value{
A plot.
}
\description{
Plot the minimization history as a function of the number of iterations.
}
\examples{
library(EmiR)

x1 <- parameter("x1", -32.768, 32.768)
x2 <- parameter("x2", -32.768, 32.768)
l <- list(x1, x2)

config <- config_ps(iterations = 250, n_particles = 150)
ps <- minimize_ps(obj_func = ackley_func,
                  parameters = l,
                  config = config)
plot_history(ps)
}
