% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_history.R
\name{plot_population}
\alias{plot_population}
\title{Plot population}
\usage{
plot_population(minimizer_result, iteration, n_points = 1000, ...)
}
\arguments{
\item{minimizer_result}{an object of class \code{MinimizationResult}.}

\item{iteration}{iteration.}

\item{...}{additional arguments, such as graphical parameters (see \link[graphics]{contour}).}
}
\value{
A plot.
}
\description{
Plot the XXX.
}
\examples{
library(EmiR)

x1 <- parameter("x1", -32.768, 32.768)
x2 <- parameter("x2", -32.768, 32.768)
l <- list(x1, x2)

config <- config_ps(iterations = 250, n_particles = 150)
ps <- minimize_ps(obj_func = ackley_func,
                  parameters = l,
                  config = config)
plot_history(ps)
}
