% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC__Qing.R
\name{qing_func}
\alias{qing_func}
\title{Qing Function}
\usage{
qing_func(x)
}
\arguments{
\item{x}{numeric or complex vector.}
}
\value{
The value of the function.
}
\description{
\loadmathjax
Implementation of n-dimensional Qing function.
}
\details{
On an n-dimensional domain it is defined by

\mjdeqn{f(\vec{x}) = \sum_{i=1}^{n} \left( x_{i}^2 - i \right)^2,}{f(x) = sum_1^n ( x_i^2 - i )^2,}
and is usually evaluated on
\mjeqn{x_{i} \in [ -500, 500 ]}{x_{i} in [-500, 500]}, for all
\mjeqn{i=1,...,n}{i=1,...,n}. The function has global minima at
\mjeqn{f(\vec{x}) = 0}{f(x) = 0} for \mjeqn{x_{i}=\pm \sqrt{i}}{x_i = +-sqrt(i)} for all \mjeqn{i=1,...,n}{i=1,...,n}.
}
\references{
\insertRef{Qing2006}{EmiR}
}
