% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotableFunctions.R
\name{rastrigin_func}
\alias{rastrigin_func}
\title{Rastrigin function}
\usage{
rastrigin_func(x)
}
\value{

}
\description{
Implementation of n-dimensional Rastrigin function.
}
\details{
\loadmathjax
On an n-dimensional domain it is defined by:

\mjdeqn{20n + \sum_{i=1}^n \left( x_{i}^2 - 20\cos(2\pi x_{i}) \right),}{20n + sum_{i=0}^n \link[=x_{i}^2 - 20cos(2*pi*x_{i})]{x_{i}^2 - 20cos(2\emph{pi}x_{i})},}
where is usually evaluated on the hypercube
\mjeqn{x_{i} \in [ -5.12, 5.12 ]}{x_{i} in \link{-5.12, 5.12}}, for all
\mjeqn{i=1,...,n}{i=1,...,n}. The global minimum is
\mjeqn{f(0,0,...,0)=0}{f(0,0,...,0)=0}.
}
\examples{
library(EmiR)

x1 <- parameter("x1", -5.12, 5.12)
x2 <- parameter("x2", -5.12, 5.12)
x3 <- parameter("x3", -5.12, 5.12)
l <- list(x1, x2, x3)

config <- config_PS(iterations = 250, n_particles = 150)
ps <- minimize_PS(cost_function = rastrigin_func,
                  parameters = l,
                  config = config)
print(ps)
}
