% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styblinski-Tank.R
\name{styblinski_tank_func}
\alias{styblinski_tank_func}
\title{Styblinski-Tank function}
\usage{
styblinski_tank_func(x)
}
\arguments{
\item{x}{numeric or complex vector.}
}
\value{
The value of the function.
}
\description{
\loadmathjax
Implementation of n-dimensional Schwefel function.
}
\details{
On an n-dimensional domain it is defined by

\mjdeqn{f(\vec{x}) = \frac{1}{2} \sum_{i=1}^{n} \left( x_{i}^4 - 16x_{i}^2 + 5x_{i} \right),}{f(x) = 1/2 sum_1^n ( x_i^4 -16\emph{x_i^2 +5}x_i ),}
and is usually evaluated on
\mjeqn{x_{i} \in [ -5, 5 ]}{x_{i} in \link{-5, 5}}, for all
\mjeqn{i=1,...,n}{i=1,...,n}. The function has one global minimum at
\mjeqn{f(\vec{x}) = -39.16599n}{f(x) = -39.16599n} for \mjeqn{x_{i}=-2.903534}{x_i=-2.903534} for all \mjeqn{i=1,...,n}{i=1,...,n}.
}
\examples{
library(EmiR)

x1 <- parameter("x1", -500, 500)
x2 <- parameter("x2", -500, 500)
l <- list(x1, x2)

config <- config_PS(iterations = 250, n_particles = 150)
ps <- minimize_PS(cost_function = schwefel_func,
                  parameters = l,
                  config = config)
print(ps)
}
\references{
\insertRef{Styblinski1990}{EmiR}
}
