% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Plot_history.R
\name{animate_population}
\alias{animate_population}
\title{Animation of population motion}
\usage{
animate_population(minimizer_result, n_points = 100)
}
\arguments{
\item{minimizer_result}{an object of class \code{OptimizationResults} (see \link[EmiR]{OptimizationResults}).}

\item{n_points}{number of points per dimension used to draw the objective function. Default is \code{100}.}
}
\description{
Create an animation of the population motion for the minimization of
1D and 2D functions. The animation can be produced only if \code{save_pop_history} is \code{TRUE}
in the options of the minimizer (see \link[EmiR]{MinimizerOpts}).
}
