% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__ABC.R
\name{config_abc}
\alias{config_abc}
\title{Configuration object for the Artificial Bee Colony Algorithm}
\usage{
config_abc(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  absolute_tol = NULL,
  employed_frac = 0.5,
  n_scout = 1
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of bees.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the \emph{same}
(see the parameter \code{absolute_tol}) best cost before ending the minimization. If \code{NULL} the
minimization continues for the number of iterations specified by the parameter \code{iterations}.
Default is \code{NULL}.}

\item{absolute_tol}{absolute tolerance when comparing best costs from consecutive iterations.
If \code{NULL} the machine epsilon is used. Default is \code{NULL}.}

\item{employed_frac}{fraction employed bees. Default is \code{0.5}.}

\item{n_scout}{number of scout bees. Default is \code{1}.}
}
\value{
\code{config_abc} returns an object of class \code{ABCConfig}.
}
\description{
Create a configuration object for the Artificial Bee Colony Algorithm (ABC). At minimum the number of iterations
(parameter \code{iterations}) and the number of bees (parameter \code{population_size}) have
to be provided.
}
\examples{
conf <- config_abc(iterations = 100, population_size = 50, iterations_same_cost = NULL, absolute_tol = NULL, employed_frac = 0.5, n_scout = 1)

}
\references{
\insertRef{Karaboga2007}{EmiR}
}
