% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__CS.R
\name{config_cs}
\alias{config_cs}
\title{Create a configuration object for the CS minimizer}
\usage{
config_cs(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  mutation_rate = 0.25,
  penalty_parameter = 10
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{population_size}{number of ???}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{mutation_rate}{???.}

\item{penalty_parameter}{penalty parameter in constrained optimization.}
}
\value{
\code{config_cs} returns an object of class \code{CSConfig}.
}
\description{
Create a configuration object for the CS minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the population size (parameter \code{population_size}) have
to be provided.
}
