% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__PS.R
\name{config_ps}
\alias{config_ps}
\title{Create a configuration object for the PS minimizer}
\usage{
config_ps(
  iterations,
  population_size,
  iterations_same_cost = NULL,
  cognitive = 2,
  social = 2,
  inertia = 0.4,
  max_velocity = 0.1
)
}
\arguments{
\item{iterations}{maximum number of iterations.}

\item{iterations_same_cost}{maximum number of consecutive iterations with the same cost before
ending the minimization. If \code{NULL} the minimization continues for the number of iterations
specified by the parameter \code{iterations}.}

\item{cognitive}{cognitive learning factor.}

\item{social}{social learning factor.}

\item{inertia}{inertia factor.}

\item{max_velocity}{a measure of the maximum distance particles travel at each iteration.}

\item{n_particles}{number of particles.}
}
\value{
\code{config_ps} returns an object of class \code{PSConfig}.
}
\description{
Create a configuration object for the PS minimizer. At minimum the number of iterations
(parameter \code{iterations}) and the number of particles (parameter \code{n_particles}) have
to be provided.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_ps(iterations = 250, n_particles = 100)
ps <- minimize_ps(obj_func = eggholder,
                  parameters = l,
                  config = config)
print(ps)
}
