% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RInterface_CS.R
\name{minimize_cs}
\alias{minimize_cs}
\title{Cuckoo Search constrained minimization}
\usage{
minimize_cs(cost_function, constraints = NULL, parameters, config)
}
\arguments{
\item{cost_function}{cost function to be minimized.}

\item{parameters}{a list of objects of class \code{Parameter} the cost function is minimized with respect to.
See \link[EmiR]{parameter}.}

\item{config}{an object of class \code{CSConfig} with the configuration parameters
for the CS algorithm. See \link[EmiR]{config_cs}.}
}
\value{
\code{minimize_cs} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize a cost function, subjected to one or more constraints, using
the Cuckoo Search (CS) algorithm.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_cs(iterations = 250, population_size = 100)
cs <- minimize_cs(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(cs)
}
