% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS_Interface.R
\name{minimize_cs}
\alias{minimize_cs}
\title{Cuckoo Search function minimization}
\usage{
minimize_cs(obj_func, constraints = NULL, parameters, config)
}
\arguments{
\item{obj_func}{objective function be minimized.}

\item{constraints}{list of constraints. Constraints are requested to be objects of
class \code{Constraint} (see \link[EmiR]{constraint}).}

\item{parameters}{list of parameters the function is minimized with respect to.
Parmeters are requested to be objects of class \code{Parameter}(see \link[EmiR]{parameter}).}

\item{config}{an object of class \code{CSConfig} with the configuration parameters
used by algorithm (see \link[EmiR]{config_cs}).}
}
\value{
\code{minimize_cs} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize an objective function, possibly subjected to inequality constraints, using
the Cuckoo Search (CS) algorithm.
}
\details{
XXXIn case of a constrained optimization only inequality constraints are allowed. The
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_cs(iterations = 250, population_size = 100)
cs <- minimize_cs(obj_func = eggholder,
                  parameters = l,
                  config = config)
print(cs)
}
\references{
\insertRef{Yang2009}{EmiR}
}
