% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minimize_GA}
\alias{minimize_GA}
\title{Genetic Algorithm minimization}
\usage{
minimize_GA(cost_function, parameters, config)
}
\arguments{
\item{cost_function}{cost function to be minimized.}

\item{parameters}{a list of objects of class \code{Parameter} the cost function is minimized with respect to.
See \link[EmiR]{parameter}.}

\item{config}{an object of class \code{GAConfig} with the configuration parameters
for the GA algorithm. See \link[EmiR]{config_GA}.}
}
\value{
\code{minimize_GA} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize a cost function using the Genetic Algorithm (GA) algorithm.
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_GA(iterations = 250, population_size = 100)
ga <- minimize_GA(cost_function = eggholder,
                  parameters = l,
                  config = config)
print(ga)
}
