% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALGO__PS.R
\name{minimize_ps}
\alias{minimize_ps}
\title{Particle Swarm function minimization}
\usage{
minimize_ps(obj_func, parameters, config, constraints = NULL, ...)
}
\arguments{
\item{obj_func}{objective function be minimized.}

\item{parameters}{list of parameters the function is minimized with respect to.
Parmeters are requested to be objects of class \code{Parameter}(see \link[EmiR]{parameter}).}

\item{config}{an object of class \code{PSConfig} with the configuration parameters
used by algorithm (see \link[EmiR]{config_ps}).}

\item{constraints}{list of constraints. Constraints are requested to be objects of
class \code{Constraint} (see \link[EmiR]{constraint}).}
}
\value{
\code{minimize_ps} returns an object of class \code{MinimizationResult}.
}
\description{
Minimize an objective function, possibly subjected to inequality constraints, using
the Particle Swarm (PS) algorithm.
}
\details{
XXXIn case of a constrained optimization only inequality constraints are allowed. The
}
\examples{
library(EmiR)

eggholder <- function(x) {
  value = -(x[2]+47)*sin(sqrt(abs(x[1]/2+x[2]+47)))+x[1]*sin(sqrt(abs(x[1]-(x[2]+47))));
  return(value)
}

x1 <- parameter("x1", -512, 512)
x2 <- parameter("x2", -512, 512)
l <- list(x1, x2)

config <- config_ps(iterations = 250, n_particles = 100)
ps <- minimize_ps(obj_func = eggholder,
                  parameters = l,
                  config = config)
print(ps)
}
\references{
\insertRef{eberhart1995new}{EmiR}
}
