% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS__MinimizerOpts.R
\docType{class}
\name{MinimizerOpts}
\alias{MinimizerOpts}
\title{Options for EmiR optimization algorithms}
\arguments{
\item{silent_mode}{if \code{TRUE} no output to console is generated. Default is \code{FALSE}.}

\item{save_population}{if \code{TRUE} the position of all individuals in the population at
each iteration is stored. This is necessary for \code{plot_population} and \code{animate_population}
functions to work, but it should not be used for optimizations involving large populations and
large numbers of iterations. Default is \code{FALSE}.}

\item{oob_solutions}{strategy to treat out-of-bound solutions. Possible values are:
\itemize{
\item \code{"RBC"} - Reflective Boundary Condition: the solution is placed back inside the search domain
at a position which is distanced from the boundary as the out-of-bound excess. Depending on the
optimization algorithm, the velocity of the corresponding individual of the population could be
also inverted.
\item \code{"PBC"} - Periodic Boundary Condition: the solution is placed back inside the search domain
at a position which is distanced from the \emph{opposite} boundary as the out-of-bound excess.
\item \code{"BAB"} - Back At Boundary: the solution is placed back at the boundaries for the out-of-bound
dimensions.
\item \code{"DIS"} - Disregard the solution: the solution is replaced by a new one, which is randomly
generated in the search space.
}

Default is \code{"DIS"}.}
}
\description{
Options for EmiR optimization algorithms.
}
