% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Plot_history.R
\name{plot_history}
\alias{plot_history}
\title{Plot minimization history}
\usage{
plot_history(minimizer_result, ...)
}
\arguments{
\item{minimizer_result}{an object of class \code{OptimizationResults} (see \link[EmiR]{OptimizationResults}).}

\item{...}{additional arguments, such as graphical parameters (see \link[graphics]{plot}).}
}
\description{
Plot the minimization history as a function of the number of iterations.
}
