% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METHOD__Plot_history.R
\name{plot_population}
\alias{plot_population}
\title{Plot the population position}
\usage{
plot_population(minimizer_result, iteration, n_points = 100)
}
\arguments{
\item{minimizer_result}{an object of class \code{OptimizationResults} (see \link[EmiR]{OptimizationResults}).}

\item{iteration}{iteration at which the population is plotted.}

\item{n_points}{number of points per dimention used to draw the objective function. Default is \code{100}.}
}
\description{
Plot the position of all individuals in the population, at a given iteration, for 1D and 2D functions. The plot can be
produced only if \code{save_pop_history} is \code{TRUE} in the options of the minimizer (see \link[EmiR]{MinimizerOpts}).
}
